pdf-search/requirements.txt
# Pinecone SDK
pinecone-client>=3.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0  # Alternative PDF extractor with better text extraction

# Text Processing
langchain>=0.1.0  # For text splitting and chunking
langchain-text-splitters>=0.0.1

# HTTP requests (for fetching PDFs from Google Drive)
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For better chunking
tiktoken>=0.5.0  # Token counting

# Optional: For reranking
sentence-transformers>=2.2.0  # If using local reranking

# Utilities
tqdm>=4.66.0  # Progress bars
